api-art-analytics/app/requirements.txt
gunicorn==22.0.0
flask==3.0.3
flask_cors==4.0.1
flask_limiter==3.7.0
flask_jwt_extended==4.6.0
flask_praetorian==1.6.0
flask_sqlalchemy==3.1.1
flask_mail==0.10.0
omegaconf==2.3.0
pydantic==2.8.2
python-dotenv==1.0.1
psycopg2===2.9.9
pandas==2.2.2
pgvector==0.3.0
sqlparse==0.5.0
numpy==1.26.4
